Move API functions to a separate file
